refact:  extract setPolicy into utils
